WIP
Added Utils module.
New enhanced variant of expandReferences, called 'finalizeOperands'.
Still in the making, but looking promising is 'wrapOperation',
which handles the finalization of arguments and the execution of
a operation.

New term introduced, `Selector`: a function that takes state, but does
not return it. Which gives us some clarity around the difference between
an operation (which always returns state) and the dataSource functions
(which need state but don't return it).